requirements.txt
pytesseract
opencv-python
numpy
pyautogui
Pillow
requests
tk
keyboard
